further restructuring
move existing booth files to new booth directory
